06_Youtube_Transcript_Agent/pyproject.toml
[project]
name = "06-youtube-transcript-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.3",
    "openai>=1.12.0",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=0.6.2",
]
